Mino test cleanups
