+++ python/tests/core/test_transformation_function_engine.py
@@ -14,9 +14,7 @@
 #   limitations under the License.
 #
 
-import datetime
 
-import numpy
 import pytest
 from hsfs import (
     engine,
@@ -926,433 +924,6 @@ class TestTransformationFunctionEngine:
         # Assert
         assert transformation_fn_dict["tf_name"] != tf_attached
         assert transformation_fn_dict["tf1_name"] != tf1_attached
-
-    def test_infer_spark_type_string_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(str)
-
-        # Assert
-        assert result == "STRING"
-
-    def test_infer_spark_type_string_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("str")
-
-        # Assert
-        assert result == "STRING"
-
-    def test_infer_spark_type_string_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("string")
-
-        # Assert
-        assert result == "STRING"
-
-    def test_infer_spark_type_byte_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(bytes)
-        result1 = tf_engine.infer_spark_type("BinaryType()")
-
-        # Assert
-        assert result == "BINARY"
-        assert result1 == "BINARY"
-
-    def test_infer_spark_type_int8_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.int8)
-
-        # Assert
-        assert result == "BYTE"
-
-    def test_infer_spark_type_int8_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("int8")
-
-        # Assert
-        assert result == "BYTE"
-
-    def test_infer_spark_type_int8_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("byte")
-        result1 = tf_engine.infer_spark_type("ByteType()")
-
-        # Assert
-        assert result == "BYTE"
-        assert result1 == "BYTE"
-
-    def test_infer_spark_type_int16_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.int16)
-
-        # Assert
-        assert result == "SHORT"
-
-    def test_infer_spark_type_int16_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("int16")
-
-        # Assert
-        assert result == "SHORT"
-
-    def test_infer_spark_type_int16_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("short")
-        result1 = tf_engine.infer_spark_type("ShortType()")
-
-        # Assert
-        assert result == "SHORT"
-        assert result1 == "SHORT"
-
-    def test_infer_spark_type_int_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(int)
-
-        # Assert
-        assert result == "INT"
-
-    def test_infer_spark_type_int_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("int")
-
-        # Assert
-        assert result == "INT"
-
-    def test_infer_spark_type_int_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.int32)
-        result1 = tf_engine.infer_spark_type("IntegerType()")
-
-        # Assert
-        assert result == "INT"
-        assert result1 == "INT"
-
-    def test_infer_spark_type_int64_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.int64)
-
-        # Assert
-        assert result == "LONG"
-
-    def test_infer_spark_type_int64_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("int64")
-
-        # Assert
-        assert result == "LONG"
-
-    def test_infer_spark_type_int64_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("long")
-
-        # Assert
-        assert result == "LONG"
-
-    def test_infer_spark_type_int64_type_4(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("bigint")
-        result1 = tf_engine.infer_spark_type("LongType()")
-
-        # Assert
-        assert result == "LONG"
-        assert result1 == "LONG"
-
-    def test_infer_spark_type_float_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(float)
-
-        # Assert
-        assert result == "FLOAT"
-
-    def test_infer_spark_type_float_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("float")
-        result1 = tf_engine.infer_spark_type("FloatType()")
-
-        # Assert
-        assert result == "FLOAT"
-        assert result1 == "FLOAT"
-
-    def test_infer_spark_type_double_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.float64)
-
-        # Assert
-        assert result == "DOUBLE"
-
-    def test_infer_spark_type_double_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("float64")
-
-        # Assert
-        assert result == "DOUBLE"
-
-    def test_infer_spark_type_double_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("double")
-        result1 = tf_engine.infer_spark_type("DoubleType()")
-
-        # Assert
-        assert result == "DOUBLE"
-        assert result1 == "DOUBLE"
-
-    def test_infer_spark_type_timestamp_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(datetime.datetime)
-
-        # Assert
-        assert result == "TIMESTAMP"
-
-    def test_infer_spark_type_timestamp_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(numpy.datetime64)
-        result1 = tf_engine.infer_spark_type("TimestampType()")
-
-        # Assert
-        assert result == "TIMESTAMP"
-        assert result1 == "TIMESTAMP"
-
-    def test_infer_spark_type_date_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(datetime.date)
-        result1 = tf_engine.infer_spark_type("DateType()")
-
-        # Assert
-        assert result == "DATE"
-        assert result1 == "DATE"
-
-    def test_infer_spark_type_bool_type_1(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type(bool)
-
-        # Assert
-        assert result == "BOOLEAN"
-
-    def test_infer_spark_type_bool_type_2(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("boolean")
-
-        # Assert
-        assert result == "BOOLEAN"
-
-    def test_infer_spark_type_bool_type_3(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        result = tf_engine.infer_spark_type("bool")
-        result1 = tf_engine.infer_spark_type("BooleanType()")
-
-        # Assert
-        assert result == "BOOLEAN"
-        assert result1 == "BOOLEAN"
-
-    def test_infer_spark_type_wrong_type(self):
-        # Arrange
-        feature_store_id = 99
-
-        tf_engine = transformation_function_engine.TransformationFunctionEngine(
-            feature_store_id
-        )
-
-        # Act
-        with pytest.raises(TypeError) as e_info:
-            tf_engine.infer_spark_type("wrong")
-
-        # Assert
-        assert str(e_info.value) == "Not supported type wrong."
 
     def test_compute_transformation_fn_statistics(self, mocker):
         # Arrange
